cumulus_primitives_utility

Trait ChargeWeightInFungibles

pub trait ChargeWeightInFungibles<AccountId, Assets: Inspect<AccountId>> {
    // Required method
    fn charge_weight_in_fungibles(
        asset_id: <Assets as Inspect<AccountId>>::AssetId,
        weight: Weight,
    ) -> Result<<Assets as Inspect<AccountId>>::Balance, XcmError>;
}

ChargeWeightInFungibles trait, which converts a given amount of weight and an assetId, and it returns the balance amount that should be charged in such assetId for that amount of weight

Required Methods

fn charge_weight_in_fungibles( asset_id: <Assets as Inspect<AccountId>>::AssetId, weight: Weight, ) -> Result<<Assets as Inspect<AccountId>>::Balance, XcmError>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.