Crate cumulus_primitives_utility

Helper datatypes for cumulus. This includes the ParentAsUmp routing type which will route messages into an UpwardMessageSender if the destination is Parent.

Structs

ParentAsUmp

Xcm router which recognises the Parent destination and handles it by sending the message into the given UMP UpwardMessageSender implementation. Thus this essentially adapts an UpwardMessageSender trait impl into a SendXcm trait impl.