Trait cumulus_primitives_utility::ChargeWeightInFungibles

pub trait ChargeWeightInFungibles<AccountId, Assets: Inspect<AccountId>> {
    fn charge_weight_in_fungibles(
        asset_id: <Assets as Inspect<AccountId>>::AssetId, 
        weight: Weight
    ) -> Result<<Assets as Inspect<AccountId>>::Balance, XcmError>;
}

ChargeWeightInFungibles trait, which converts a given amount of weight and an assetId, and it returns the balance amount that should be charged in such assetId for that amount of weight

Required Methods

fn charge_weight_in_fungibles(
asset_id: <Assets as Inspect<AccountId>>::AssetId,
weight: Weight
) -> Result<<Assets as Inspect<AccountId>>::Balance, XcmError>